// src/descriptions/LocationDescription.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface AreaDescription {
  links?: string[];
  shape?: {
    rows: number;
    cols: number;
    state?: {
      [k: string]: unknown;
    };
  };
  state?: {
    [k: string]: unknown;
  };
  tags?: string[];
}